Modularize the check code
