+++ pkg/oscheck/linux.go
@@ -2,8 +2,12 @@ package oscheck
 
 import (
 	"fmt"
+
+	"github.com/shawnstephens517/go_enumer/pkg/logging"
 )
 
-func LinCheck() {
+// LinCheck - Future exported function for Linux Checks
+func LinCheck() []logging.LogEntry {
 	fmt.Println("Gathering basic Linux info")
+	return nil
 }
