+++ pkg/oscheck/mac.go
@@ -7,6 +7,7 @@ import (
 	"fmt"
 )
 
+// MacCheck - Future exported function for Mac Checks
 func MacCheck() {
 	fmt.Println("Running Mac Enumer")
 }
